Rust upgrade to rustc 1.6.0-nightly (5b4986fa5 2015-11-08)

<s>DO NOT r+ or try+ this</s>

<s>It causes an OOM (https://github.com/rust-lang/rust/issues/29740) and can crash the OS. Probably will set our CI on fire. </s>

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/servo/servo/8446)
<!-- Reviewable:end -->

+++ ports/cef/string_list.rs
@@ -39,7 +39,7 @@ pub extern "C" fn cef_string_list_value(lt: *mut cef_string_list_t, index: c_int
         if index as usize > (*lt).len() - 1 { return 0; }
         let ref string = (*lt)[index as usize];
         let utf16_chars: Vec<u16> = Utf16Encoder::new(string.chars()).collect();
-        cef_string_utf16_set(utf16_chars.as_ptr(), utf16_chars.len() as u64, value, 1)
+        cef_string_utf16_set(utf16_chars.as_ptr(), utf16_chars.len(), value, 1)
     }
 }
